% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontsize.R
\name{geom_text_fontsize}
\alias{geom_text_fontsize}
\title{Match ggplot2 pt scale to text scale}
\usage{
geom_text_fontsize(size = 12)
}
\arguments{
\item{size}{text scale size}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function matchs ggplot2 pt scale to text scale
}
\references{
https://stackoverflow.com/questions/52304141/adjust-font-size-to-size-of-plot-device
https://stackoverflow.com/questions/17311917/ggplot2-the-unit-of-size
https://www.tidyverse.org/blog/2020/08/taking-control-of-plot-scaling/
https://www.christophenicault.com/post/understand_size_dimension_ggplot2/
}
