% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz-package.R
\docType{package}
\name{ggviz-package}
\alias{ggviz}
\alias{ggviz-package}
\title{ggviz: Tools and themes for making fancy pĺots.}
\description{
This package provides tools and themes for making fancy pĺots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arturhgq/ggviz}
  \item \url{https://arturhgq.github.io/ggviz/}
  \item \url{https://arturhgq.github.io/dviz/}
  \item Report bugs at \url{https://github.com/arturhgq/ggviz/issues}
}

}
\author{
\strong{Maintainer}: Artur Quirino \email{contact@arturhgq.com} (\href{https://orcid.org/000-0001-8071-6348}{ORCID})

}
\keyword{internal}
