% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_col.R
\name{plot_col_wrap}
\alias{plot_col_wrap}
\title{plot col wrap}
\usage{
plot_col_wrap(
  data = NULL,
  mapping = NULL,
  stat = "identity",
  position = "dodge",
  key_glyph = NULL,
  width = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  group,
  lab_y,
  ...
)
}
\description{
plot col wrap
}
\examples{
ggplot2::presidential |>
dplyr::mutate(
  days = as.numeric(end - start) + 1
) -> data

plot_col_wrap(
  data,
  mapping = ggplot2::aes(x = name, y = days),
  width = .8,
  group = c(party),
  size = .1,
  fill = "gray"
)
}
