% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R
\name{theme_academic}
\alias{theme_academic}
\title{ggplot2 academic theme}
\usage{
theme_academic(
  color_text = "black",
  hjust_axis_title_x = 1,
  vjust_axis_title_y = 1,
  color_axis_title = "black",
  ...
)
}
\description{
ggplot2 academic theme
}
