% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R, R/theme_basic.R
\name{theme_academic}
\alias{theme_academic}
\alias{theme_basic}
\title{ggplot2 academic theme}
\usage{
theme_academic(
  family = "Times New Roman",
  color_text = "black",
  hjust_axis_title_x = 1,
  vjust_axis_title_y = 1,
  color_axis_title = "black",
  margin_axis_title_x_b = 0,
  size_grid_major_line = 0.3,
  margin_legend_t = -20,
  pos_legend = "bottom",
  panel_spacing = 1,
  color_ticks_x = "black",
  color_line_x = "black",
  length_ticks = 0.1,
  ...
)

theme_basic(
  family = "",
  size = 12,
  size_title = 12,
  size_subtitle = 10,
  size_caption = 10,
  size_strip = 10,
  size_axis = 10,
  size_axis_title = 12,
  size_grid_major_line = 0.5,
  size_legend = 10,
  margin_strip_t = 2,
  margin_strip_r = 0,
  margin_strip_b = 2,
  margin_strip_l = 0,
  margin_plot_t = 5,
  margin_plot_r = 5,
  margin_plot_b = 5,
  margin_plot_l = 5,
  margin_legend_t = 0,
  margin_legend_r = 0,
  margin_legend_b = 0,
  margin_legend_l = 0,
  margin_axis_title_x_b = 5,
  color_text = "gray30",
  color_axis_title = "gray30",
  color_ticks_x = NA,
  color_line_x = NA,
  length_ticks = 0,
  hjust_axis_title_x = 0.5,
  vjust_axis_title_y = 0.5,
  angle_axis_title_y = 90,
  pos_legend = "top",
  panel_spacing = 0.4
)
}
\description{
ggplot2 academic theme

ggplot2 theme basic
}
