% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_facet}
\alias{theme_facet}
\title{ggplot2 minimal facet theme}
\usage{
theme_facet(
  family = "",
  size = 12,
  size_title = 12,
  size_subtitle = 10,
  size_caption = 10
)
}
\description{
ggplot2 minimal facet theme
}
