% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontsize.R
\name{geom_text_fontsize}
\alias{geom_text_fontsize}
\title{Adequa a escala de pontos do ggplot2 à escala de texto}
\usage{
geom_text_fontsize(.size = 12)
}
\description{
Adequa a escala de pontos do ggplot2 à escala de texto
}
\references{
https://stackoverflow.com/questions/52304141/adjust-font-size-to-size-of-plot-device
https://stackoverflow.com/questions/17311917/ggplot2-the-unit-of-size
https://www.tidyverse.org/blog/2020/08/taking-control-of-plot-scaling/
https://www.christophenicault.com/post/understand_size_dimension_ggplot2/
}
