% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_basic.R
\name{theme_basic}
\alias{theme_basic}
\title{ggplot2 theme basic}
\usage{
theme_basic(
  family = "",
  size = 12,
  size_title = 12,
  size_subtitle = 10,
  size_caption = 10,
  size_strip = 10,
  size_axis = 10,
  size_axis_title = 12,
  strip_margin_t = 2,
  strip_margin_r = 0,
  strip_margin_b = 2,
  strip_margin_l = 0,
  plot_margin_t = 5,
  plot_margin_r = 5,
  plot_margin_b = 5,
  plot_margin_l = 5,
  color_text = "gray30",
  color_axis_title = "gray30",
  hjust_axis_title_x = 0.5,
  vjust_axis_title_y = 0.5,
  angle_axis_title_y = 90
)
}
\description{
ggplot2 theme basic
}
