% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R, R/theme_basic.R,
%   R/theme_boxplot.R
\name{theme_academic}
\alias{theme_academic}
\alias{theme_basic}
\alias{theme_boxplot}
\title{ggplot2 themes}
\usage{
theme_academic(
  family = "Times New Roman",
  color_text = "black",
  hjust_axis_title_x = 1,
  vjust_axis_title_y = 1,
  color_axis_title = "black",
  margin_axis_title_x_b = 0,
  size_strip = 12,
  color_plot_background = "black",
  color_strip_background = NA,
  fill_strip_background = NA,
  halign_strip_text = 0,
  face_strip_text = "bold",
  margin_legend_t = 0,
  pos_legend = "top",
  panel_spacing = 1,
  color_ticks_x = "black",
  color_line_x = "black",
  length_ticks = 0.1,
  ...
)

theme_basic(
  family = "",
  size = 12,
  size_title = 12,
  size_subtitle = 10,
  size_caption = 10,
  size_strip = 10,
  size_axis_text = 10,
  size_axis_title = 12,
  size_grid_major_line_y = 0.5,
  size_grid_major_line_x = NA,
  size_grid_minor_line_y = NA,
  size_grid_minor_line_x = NA,
  size_legend = 10,
  size_ticks = 0.2,
  size_line = 0.2,
  margin_strip_t = 2,
  margin_strip_r = 0,
  margin_strip_b = 0.5,
  margin_strip_l = 0,
  padding_strip_t = 0,
  padding_strip_r = 0,
  padding_strip_b = 0,
  padding_strip_l = 0,
  margin_plot_t = 5,
  margin_plot_r = 5,
  margin_plot_b = 5,
  margin_plot_l = 5,
  margin_legend_t = 4,
  margin_legend_r = 4,
  margin_legend_b = 2,
  margin_legend_l = 4,
  margin_axis_title_x_b = 5,
  color_text = "gray30",
  color_axis_title = "gray30",
  color_ticks_x = NA,
  color_line_x = NA,
  color_grid = color("linegray"),
  color_plot_background = "white",
  color_strip_background = "grey86",
  fill_strip_background = "grey86",
  halign_strip_text = 0.5,
  face_strip_text = "plain",
  length_ticks = 0,
  hjust_axis_title_x = 0.5,
  vjust_axis_title_y = 0.5,
  angle_axis_title_y = 90,
  pos_legend = "bottom",
  panel_spacing = 0.4,
  linetype_grid = "solid",
  ...
)

theme_boxplot(
  family = "Times New Roman",
  color_text = "black",
  hjust_axis_title_x = 1,
  vjust_axis_title_y = 1,
  color_grid = "darkgray",
  color_axis_title = "black",
  color_plot_background = "black",
  margin_axis_title_x_b = 0,
  color_strip_background = NA,
  fill_strip_background = NA,
  halign_strip_text = 0,
  margin_legend_t = 0,
  margin_strip_t = 5,
  margin_strip_r = 0,
  margin_strip_b = 2,
  margin_strip_l = 0,
  padding_strip_t = 5,
  padding_strip_r = 0,
  padding_strip_b = 0,
  padding_strip_l = 0,
  size_grid_major_line_y = NA,
  size_grid_major_line_x = 0.1,
  size_ticks = NA,
  size_line = NA,
  pos_legend = "top",
  panel_spacing = 1,
  color_ticks_x = "black",
  color_line_x = "black",
  length_ticks = 0.1,
  linetype_grid = "dashed",
  ...
)
}
\arguments{
\item{...}{theme_basic arguments}
}
\description{
ggplot2 themes
}
