% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_facet.R
\name{theme_facet}
\alias{theme_facet}
\title{ggplot2 facet theme}
\usage{
theme_facet(design = "basic", ...)
}
\arguments{
\item{design}{"basic" or ""academic"}
}
\description{
ggplot2 facet theme
}
\examples{
dplyr::starwars |>
  tidyr::drop_na(eye_color, height, mass) |>
  dplyr::filter(mass < 1200) |>
  dplyr::mutate(height = height/100) |>
  ggplot2::ggplot() +
  ggplot2::facet_wrap(~ eye_color) +
  ggplot2::aes(x = mass, y = height, color = "#333333", fill  = "#3399ff") +
  ggplot2::geom_point(shape = 21, stroke = 0.3, alpha = .7, size = 3) +
  ggplot2::scale_y_continuous(
    labels = scales::label_comma(decimal.mark = ",", big.mark = "."),
    limits = c(0, 3)
  ) +
  ggplot2::scale_x_continuous(limits = c(0, 160), breaks = seq(0, 160, 30)) +
  ggplot2::scale_color_identity() +
  ggplot2::scale_fill_identity() +
  ggviz::theme_facet() +
  ggplot2::labs(
    y = "altura (metros)",
    x = "massa (kg)",
    title = "Altura e massa dos personagens de Star Wars por cor dos olhos",
    subtitle = "Aproximadamente 24\% dos personagens tem olhos marrons e 9,2\% olhos laranjas.\nO personagem mais alto da franquia é o Yarael Poof, com 2,64 metros.\nJabba, por sua vez, é o mais pesado, com 1.358 quilos.",
    caption = "Como o peso do Jabba destoa dos demais, para melhorar a visualização dos dados, optei por excluí-lo.\nDados relativos aos 7 primeiros filmes da franquia."
  ) +
  ggplot2::coord_cartesian(clip = "off", expand = TRUE)
}
